right cel highligh

+++ Assets/Scripts/Game/Logic/Definitions/Constructions/ConstructionsSettingsDefinition.cs
@@ -2,7 +2,7 @@
 {
     public class ConstructionsSettingsDefinition
     {
-        public int CellSize { get; set; }
+        public float CellSize { get; set; }
     }
 }
